Required Questions - Make input checks more robust
For required questions, the function `checkIndividual()` returns TRUE if there is an input is not NULL, not NA, and not equal to "". Most, if not all, shiny inputs should return one of these values if they have no data. When checking for equality with "", input types that cannot be coerced to class character throws an error described in #32.

Explicitly coercing the input values to class character when performing this check, we should avoid this error. This is especially relevant with custom [date inputs](https://www.jdtrat.com/blog/extending-shinysurveys/#adding-a-dateinput).